Fix event priority and with it 'unhide code' an event listener.
After take a look at the event net, I realize that `Hal` renderer is attached with priority 100, and as the whole point to attach a listener to the event `MvcEvent::EVENT_RENDER_ERROR` is catch any rendering error this listener attachment ('ZF\ApiProblem\RenderErrorListener' service) must happen before the `Hal`'s renderer listener get executed. I pick 400 as priority value becuase is important that this listener be executed after the `ApiProblemListener::onRender` listener, and the priority of that listener is 1000.