Created Application extension for catching route exceptions
This patch provides a new class, `ZF\Apigility\Application`, which
extends `Zend\Mvc\Application` to provide a try/catch block around the
route event. This addresses the concerns raised in #83, as it means any
exception raised by a route listener will be caught and lead to
triggering the `dispatch.error` event. In the case of Apigility
applications, this will allow us to return problem details.